+++ tests/resources/demo/__init__.py
@@ -1,3 +1,3 @@
-#This file is generated for the module 'demo' on Mon Dec 14 22:26:54 2020.
+#This file is generated for the module 'demo' on Tue Dec 15 14:46:32 2020.
 
 from . import *