+++ tests/resources/demo/demo.py
@@ -1,6 +1,6 @@
 """
 The python file contains the different classes for accessing the containers and parameters of the 'demo' module.
-This file is generated for the module 'demo' on Mon Dec 14 22:26:54 2020.
+This file is generated for the module 'demo' on Tue Dec 15 14:46:32 2020.
 """
 
 from lxml import etree
@@ -639,10 +639,5 @@ class contC(AutosarNode):
 # Exception classes
 class ValueNotPossibleError(Exception):
     """Exception raised when the provided value is not possible to set for the parameter"""
-    def __init__(self, message):
-        super().__init__(message)
-
-class FileSaveError(Exception):
-    """Exception raised when the file is unable to save"""
     def __init__(self, message):
         super().__init__(message)